Removed end: present for everybody except the handful still at the EF.
This leaves people who were never part of the EF in a somewhere weird position, with ??? end-date even though they may still be active in the ecosystem.

We need more data at reflect these different scenarios.

Start and end in Ethereum ecosystem
Start and end in the EF

And maybe we even want to start adding people from ConsenSys or other projects, and then start filtering on those as categories?